Project/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
paho-mqtt==2.1.0
websockets==13.1
